refactor: 파일의 기본 경로도 설정하여 사용할 수 있도록 변경
