+++ src/main/java/codesquad/server/router/RouteEntryManager.java
@@ -1,22 +1,23 @@
 package codesquad.server.router;
 
-import codesquad.app.usecase.UserUsecase;
 import codesquad.container.Component;
-import codesquad.container.ContainerHolder;
-import codesquad.http.HttpMethod;
 
 import java.util.ArrayList;
 import java.util.List;
 
 @Component
 public class RouteEntryManager {
-    private final List<RouteEntry> routeEntry;
+    private final List<RouteEntry> routeEntries = new ArrayList<>();
 
-    public RouteEntryManager(RouteEntryConfigurer routeEntryConfigurer) {
-        routeEntry = routeEntryConfigurer.getRouteEntries();
+    public RouteEntryManager() {
     }
 
-    public List<RouteEntry> getRouteEntry() {
-        return routeEntry;
+    public RouteEntryManager add(RouteEntry routeEntry) {
+        routeEntries.add(routeEntry);
+        return this;
+    }
+
+    public List<RouteEntry> getRouteEntries() {
+        return routeEntries;
     }
 }
